Move postgres database utility classes to TestUtils.
