+++ src/SqlStreamStore.TestUtils/SqlStreamStore.TestUtils.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="Docker.DotNet" Version="3.125.2" />
+    <PackageReference Include="Npgsql" Version="4.0.2" />
     <PackageReference Include="Serilog" Version="2.7.1" />
     <PackageReference Include="Serilog.Sinks.Observable" Version="2.0.2" />
     <PackageReference Include="System.Data.SqlClient" Version="4.6.0" />
